+++ src/steamship/agents/example_agents/my_assistant.py
@@ -2,11 +2,11 @@ from typing import List, Optional
 
 from steamship import Block
 from steamship.agents.base import Metadata
+from steamship.agents.communication.chat_message import ChatMessage
 from steamship.agents.planner.react import OpenAIReACTPlanner
 from steamship.agents.service.agent_service import AgentService
 from steamship.agents.tools.image_generation.generate_image import GenerateImageTool
 from steamship.agents.tools.search.search import SearchTool
-from steamship.experimental.transports.chat import ChatMessage
 
 # AgentService is a PackageService.
 from steamship.utils.repl import AgentREPL
