+++ src/steamship/experimental/transports/steamship_widget.py
@@ -2,7 +2,7 @@ import uuid
 from typing import Optional
 
 from steamship import SteamshipError
-from steamship.experimental.transports.chat import ChatMessage
+from steamship.agents.communication.chat_message import ChatMessage
 from steamship.experimental.transports.transport import Transport
 
 API_BASE = "https://api.telegram.org/bot"
